% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBaseRoadDvmt.R
\name{CalculateBaseRoadDvmt}
\alias{CalculateBaseRoadDvmt}
\title{Assign base year DVMT by vehicle type and road class for Mareas}
\usage{
CalculateBaseRoadDvmt(L)
}
\arguments{
\item{L}{A list containing data defined by the module specification.}
}
\value{
A list containing data produced by the function consistent with the
module specifications.
}
\description{
\code{CalculateBaseRoadDvmt} calculates base year DVMT on roadways within the
urbanized portions of Mareas by vehicle type (light-duty vehicle,
heavy truck, bus) and road class (freeway, arterial, other). It computes
some other values as explained in the details.
}
\details{
This function calculates several values necessary for calculating road
performance and vehicle energy consumption and emissions. These include
calculating base year daily vehicle miles of travel (DVMT) by vehicle type
(light-duty, heavy truck, bus), and jointly by vehicle type and road class
(freeway, arterial, other) for roadways in the urbanized proportions in
Mareas. The function also computes required parameters used for calculating
future year DVMT by vehicle type and road class. If region-level inputs have
been provided (region_base_year_hvytrk_dvmt.csv), the function computes base
year heavy truck DVMT and ratios of heavy truck DVMT to total region
population and income. The function calculates base year commercial service
light-duty vehicle (LDV) DVMT and ratios of that DVMT to population, and
income by Marea. If the user has not provided a base year input for LDV DVMT
on urbanized area roads in an Marea, that DVMT is calculated based on the
base year urbanized area population and default values of LDV DVMT per capita
for the urbanized area. The ratio of LDV urbanized area road DVMT to total
DVMT of households residing in the urbanized area, commercial service LDV
associated with resident households, and public transit LDV DVMT is
calculated for each Marea. This ratio is used to calculate future year LDV
roadway DVMT from total urbanized area LDV demand. If the user has not
provided base year urbanized area heavy truck roadway DVMT for an Marea, the
function computes that value using a default value of heavy truck DVMT per
capita for the urbanized area and the base year urbanized area population.
The ratios of urbanized area heavy truck road DVMT to urbanized area
population and urbanized area income are calculated to use in calculating
future year heavy truck DVMT. If the user has not provided inputs for the
DVMT proportions of each vehicle type by road class, the function calculates
those proportions from default data for the urbanized area. Using those DVMT
proportions, the function computes urbanized area roadway DVMT by vehicle
type and road class.
}
