% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\name{calculateCongestion}
\alias{calculateCongestion}
\title{Calculate DVMT or DVHT by congestion level.}
\usage{
calculateCongestion(MeasureType, RoadType, LaneMi, DVMT, DVHT = NULL)
}
\arguments{
\item{MeasureType}{a string having the value 'DVMT' or 'DVHT' depending on
whether daily vehicle miles of travel (DVMT) or daily vehicle hours of travel
is to be split into congestion levels.}

\item{RoadType}{a string having the value 'Fwy' or 'Art' depending on whether
congestion is to be calculated for freeways (Fwy) or arterial roadways (Art).}

\item{LaneMi}{a number identifying the number of lane-miles for the roadway
type.}

\item{DVMT}{a number identifying the daily vehicle miles of travel on the
roadway type.}

\item{DVHT}{a number identifying the daily vehicle hours of travel on the
roadway type. Note that a value is required for this argument only if the
MeasureType is 'DVHT'.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateCongestion} splits input DVMT or DVHT into the amounts in
each of 5 congestion levels:none, moderate, heavy, severe, and extreme.
}
\details{
This function splits input DVMT or DVHT into amounts by congestion level
(none, moderate, heavy, severe, and extreme) as a function of the ratio of
DVMT to lane-miles. The lookup tables in CongestedProportions_ls are used
to lookup proportions by congestion level. The input DVMT or DVHT are
split with these proportions and the split values are returned.
}
