% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{DvmtSplit_LM}
\alias{DvmtSplit_LM}
\title{DVMT split factor model}
\format{A linear model object
\describe{
  \item{DvmtSplit_LM}{a linear model object}
}}
\source{
CalculateRoadPerformance.R script.
}
\usage{
DvmtSplit_LM
}
\description{
A linear model for calculating the factor used to calculate the ratio of
freeway DVMT to arterial DVMT from the ratio of freeway average speed to
arterial average speed.
}
\keyword{datasets}
