% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRoadPerformance.R
\docType{data}
\name{UmsInp_ls}
\alias{UmsInp_ls}
\title{effective' speed for freeways and for arterials. The ratio of freeway and}
\format{An object of class \code{list} of length 5.}
\usage{
UmsInp_ls
}
\description{
effective' speed for freeways and for arterials. The ratio of freeway and
}
\keyword{datasets}
